+++ internal/drain.go
@@ -48,7 +48,6 @@ func (s *Server) matchNode(ctx context.Context, id string) (*v1.Node, error) {
 	fields := logrus.Fields{
 		"id": id,
 	}
-
 	s.log.WithFields(fields).Info("fleetlock: match Zincati request to Kubernetes node")
 
 	nodes, err := s.kubeClient.CoreV1().Nodes().List(ctx, metav1.ListOptions{})
@@ -58,9 +57,11 @@ func (s *Server) matchNode(ctx context.Context, id string) (*v1.Node, error) {
 	}
 
 	for _, node := range nodes.Items {
-		zincatiID, err := ZincatiID(node.Status.NodeInfo.SystemUUID)
+		zincatiID, err := ZincatiID(node.Status.NodeInfo.MachineID)
 		if err == nil && id == zincatiID {
 			fields["node"] = node.GetName()
+			fields["machineID"] = node.Status.NodeInfo.MachineID
+			fields["systemUUID"] = node.Status.NodeInfo.SystemUUID
 			s.log.WithFields(fields).Info("fleetlock: Zincati request matches Kubernetes node")
 			return &node, nil
 		}