+++ CHANGES.md
@@ -4,6 +4,12 @@ Notable changes between versions.
 
 ## Latest
 
+* Identify Kubelet nodes via `MachineID` instead of `SystemUUID` (**action required**) ([#96](https://github.com/poseidon/fleetlock/pull/96))
+  * Host `SystemUUID` and `MachineID` are identical on most platforms, but not on Azure or bare-metal
+  * Kubelet detected `SystemUUID` via sysfs, but `MachineID` requires Kubelet
+  be able to read `/etc/machine-id` from the host (`ro` mount when Kubelet
+  run as a container) (**action required**)
+  * Add support for Azure and bare-metal plaforms by changing to `MachineID`
 * Reset `fleet_lock_state` gauge on first `lock` or `unlock` call ([#71](https://github.com/poseidon/fleetlock/pull/71))
 * Improve reject reply when a client attempts to unlock a lock it doesn't own ([#71](https://github.com/poseidon/fleetlock/pull/71))